Equation: Refactor source files.
