+++ lib/kemuEquation/Equation.js
@@ -1,123 +1,11 @@
-const math = require('mathjs')
-
+const math         = require('mathjs')
 const mathsteps    = require('../../index.js')
+const EquationTerm = require('./EquationTerm')
+const Node         = require('../node')
 const stepThrough  = require('../simplifyExpression/stepThrough')
 const simplifyCore = require('../simplifyExpression/kemuCommonSearch/simplifyCore.js')
 
-const Node = require('../node')
-
-const NODE_ONE = Node.Creator.constant(1)
 const MAX_ITERATIONS = 20
-
-const EQUATION_TERM_TYPES = {
-  constant: 'C',
-  polynomial: 'W(x)',
-  function: 'f(x)',
-}
-
-class EquationTerm {
-  constructor(type, node, data) {
-    this.type = type
-    this.node = node
-    this.data = data
-  }
-
-  static createConstant(node) {
-    return new EquationTerm('constant', node)
-  }
-
-  static createPolynomial(node) {
-    return new EquationTerm('polynomial', node)
-  }
-
-  static createFunction(node) {
-    return new EquationTerm('function', node)
-  }
-
-  static operator2(op, a, b) {
-    const arithmeticMap = {
-      // C op C
-      'constant+constant': 'constant',
-      'constant-constant': 'constant',
-      'constant*constant': 'constant',
-      'constant/constant': 'constant',
-      'constant^constant': 'constant',
-
-      // W(x) op W(x)
-      'polynomial+polynomial': 'polynomial',
-      'polynomial-polynomial': 'polynomial',
-      'polynomial*polynomial': 'function',
-      'polynomial/polynomial': 'function',   // TODO: Homographic?
-      'polynomial^polynomial': 'function',   // x^x like.
-
-      // W(x) op C
-      'polynomial+constant': 'polynomial',
-      'polynomial-constant': 'polynomial',
-      'polynomial*constant': 'polynomial',
-      'polynomial/constant': 'polynomial',
-      'polynomial^constant': 'polynomial', // TODO: Review it.
-
-      // C op W(x)
-      'constant+polynomial': 'polynomial',
-      'constant-polynomial': 'polynomial',
-      'constant*polynomial': 'polynomial',
-      'constant/polynomial': 'function',   // TODO: Homographic?
-      'constant^polynomial': 'function',   // TODO: exponential?
-    }
-
-    const newNode = Node.Creator.operator(op, [a.node, b.node])
-    const hash    = a.type + op + b.type
-    let   newType = null
-
-    if ((a.type === 'function') || (b.type === 'function')) {
-      newType = 'function'
-    } else {
-      newType = arithmeticMap[hash]
-    }
-
-    if (newType == null) {
-      throw Error('unhandled node types:' + hash)
-    }
-
-    // TODO: Handle data?
-    // TODO: Handle polynomial degree?
-    return new EquationTerm(newType, newNode)
-  }
-
-  static add(a, b) {
-    return EquationTerm.operator2('+', a, b)
-  }
-
-  static sub(a, b) {
-    return EquationTerm.operator2('-', a, b)
-  }
-
-  static mul(a, b) {
-    return EquationTerm.operator2('*', a, b)
-  }
-
-  static div(a, b) {
-    return EquationTerm.operator2('/', a, b)
-  }
-
-  static pow(a, b) {
-    return EquationTerm.operator2('^', a, b)
-  }
-
-  static neg(a) {
-    const newNode = Node.Creator.unaryMinus(a.node)
-    return new EquationTerm(a.type, newNode)
-  }
-
-  toString() {
-    const typeToText = {
-      constant: 'C',
-      polynomial: 'W(x)',
-      function: 'f(x)',
-    }
-    return typeToText[this.type]
-  }
-}
 
 class Equation {
   constructor(leftNode, rightNode, comparator, unknownVariable) {
@@ -239,7 +127,7 @@ class Equation {
 
   applyStep(stepName, newNodes) {
     if (stepName === 'solution') {
-      this.solve(newNodes.items)
+      this.applySolution(newNodes.items)
 
     } else {
       // TODO: Handle comparator change.
@@ -256,6 +144,8 @@ class Equation {
     let iterIdx = 0
 
     while (goOn) {
+      console.log('')
+
       // Simplify both side before each step if needed.
       if (options && options.simplifyBeforeEachStepEnabled) {
         this.simplifyLeft()
@@ -288,7 +178,7 @@ class Equation {
     }
   }
 
-  solve(solutions) {
+  applySolution(solutions) {
     console.log('SOLUTION FOUND:', solutions.toString())
     this.solutions = solutions
   }
@@ -296,46 +186,6 @@ class Equation {
   isSolved() {
     return (this.solutions != null)
   }
-
 }
 
-class EquationSolveManager {
-  constructor() {
-    this.solveFunctionsMap = {}
-  }
-
-  registerSolveFunction(pattern, solveFunction) {
-    this.solveFunctionsMap[pattern] = solveFunction
-  }
-
-  solveEquation(equation) {
-    // Build equation pattern.
-    // Example: 'x^2 + 2x = -1' gives 'W(x)=C'
-    const pattern = equation.left.toString()
-                  + equation.comparator
-                  + equation.right.toString()
-
-    // Search for solve function for given pattern.
-    const solveFunction = this.solveFunctionsMap[pattern]
-
-    if (solveFunction == null) {
-      throw Error(`no solver for equation: '${pattern}'`)
-    }
-
-    // Run solve function.
-    return solveFunction(equation)
-  }
-}
-
-const equationSolveManager = new EquationSolveManager()
-
-module.exports = {
-  Equation: Equation,
-  EquationTerm: EquationTerm,
-  EquationSolveManager: equationSolveManager,
-}
-
-//function (equationObj, unknownVariable) {
-//  const equation = new Equation(equationObj.leftNode, equationObj.rightNode, '=', unknownVariable)
-//  return equation
-//}
+module.exports = Equation
