+++ lib/kemuEquation/test.js
@@ -1,83 +1,11 @@
-const simplifyCore = require('../simplifyExpression/kemuCommonSearch/simplifyCore.js')
-const Node         = require('../node')
-const stepThrough  = require('../simplifyExpression/stepThrough')
-const math         = require('mathjs')
+const Equation       = require('./Equation')
+const EquationSolver = require('./EquationSolver')
 
-//const math = require('mathjs')
-//const node = math.parse('EQ(2 x + 3, 4)')
+const equation = Equation.createFromString('2x - y = z', 'x')
 
-//const Equation = require('./index.js')
-//const equationNodes = Equation.Equation.createEquationFromString('g *( x ) = ( x - 4) ^ ( 2) - 3', '=')
-//const equation      = Equation.Analyze(equationNodes, 'g')
-
-const {Equation, EquationTerm, EquationSolveManager} = require('./EquationAnalyze')
-
-const NODE_ZERO  = Node.Creator.constant(0)
-const NODE_ONE   = Node.Creator.constant(1)
-const NODE_TRUE  = Node.Creator.symbol('true')
-const NODE_FALSE = Node.Creator.symbol('false')
-
-// Register equation solvers.
-EquationSolveManager.registerSolveFunction('C=C', (equation) => {
-  // Simplify both side before step.
-  equation.simplifyLeft()
-  equation.simplifyRight()
-
-  // Fetch LR nodes.
-  const leftNode  = equation.left.node
-  const rightNode = equation.right.node
-
-  // Try to calculate L-R difference.
-  const steps    = stepThrough(Node.Creator.operator('-', [leftNode, rightNode]))
-  const diffNode = steps[steps.length - 1].newNode
-
-  if (Node.Type.isConstant(diffNode)) {
-    // L - R gives a constant number.
-    // We can check equality directly.
-
-    if (Node.Type.isZero(diffNode)) {
-      // L - R = 0
-      // Solution is true.
-      equation.solve(NODE_TRUE)
-    } else {
-      // L - R != 0
-      // Solution is false.
-      equation.solve(NODE_FALSE)
-    }
-
-  } else {
-    // L - R is NOT a constant.
-    // Give up. We stuck WITHOUT solution.
-    // Possible improvement: Handle function nodes e.g. sin(1) = pi
-    // Possible improvement: Handle parameter dependend e.g. n=1
-  }
-})
-
-
-EquationSolveManager.registerSolveFunction('W(x)=C', (equation) => {
-  // Pool of known transformations.
-  const poolOfRules = [
-    // Common rules.
-    {l: 'EQ(fx + a1 , a2)', r: 'EQ(fx , a2 - a1)' , id: 'move_free_parameter_to_right'},
-    {l: 'EQ(fx - a1 , a2)', r: 'EQ(fx , a2 + a1)' , id: 'move_free_parameter_to_right'},
-
-    // Linear equation: W1(x) = C
-    {l: 'EQ(a1 x , a2)', r: 'EQ(x , a2 / a1)' , id: 'divide_both_sides'},
-    {l: 'EQ(x    , a1)', r: '[a1]'            , id: 'solution'},
-  ]
-
-  equation.applyRules(poolOfRules, {
-    simplifyBeforeEachStepEnabled: true
-  })
-})
-
-const equation = Equation.createFromString('2a+a=3a', 'x')
-
-EquationSolveManager.solveEquation(equation)
+EquationSolver.solveEquation(equation)
 
 /*
-
-
   const tests = [
     ['g *( x ) = ( x - 4) ^ ( 2) - 3', 'g = x - 8 + 13 / x'],
